test: split test_write_tracks atomically — extract validation to unit
Move test_write_duplicate_track_names_rejected to
tests/unit/dataset/test_write_tracks.py (port: input validation
fires before disk I/O). Rename the integration file to
test_write_tracks_e2e.py to avoid the pytest prepend-mode basename
collision (precedent: reconstruct plan).

Phase 5 tracks (broader) component.
