+++ tests/integration/dataset/test_write_tracks_e2e.py
@@ -128,35 +128,3 @@ def test_write_with_variants_and_tracks(tmp_path, vcf_dir: Path):
 
     meta = json.loads((out / "metadata.json").read_text())
     assert set(meta["samples"]) == {"NA00001", "NA00002", "NA00003"}
-
-
-def test_write_duplicate_track_names_rejected(tmp_path):
-    import pytest
-
-    bed = pl.DataFrame({"chrom": ["chr1"], "chromStart": [0], "chromEnd": [100]})
-    t1 = gvl.Table(
-        "dup",
-        pl.DataFrame(
-            {
-                "sample_id": ["s0"],
-                "chrom": ["chr1"],
-                "start": [0],
-                "end": [10],
-                "value": [1.0],
-            }
-        ),
-    )
-    t2 = gvl.Table(
-        "dup",
-        pl.DataFrame(
-            {
-                "sample_id": ["s0"],
-                "chrom": ["chr1"],
-                "start": [50],
-                "end": [60],
-                "value": [2.0],
-            }
-        ),
-    )
-    with pytest.raises(ValueError, match="[Dd]uplicate"):
-        gvl.write(path=tmp_path / "x.gvl", bed=bed, tracks=[t1, t2])
